pull out replica wrapper and scheduler into individual files
