+++ python/ray/serve/_private/default_impl.py
@@ -6,10 +6,12 @@ from ray.serve._private.cluster_node_info_cache import (
     ClusterNodeInfoCache,
     DefaultClusterNodeInfoCache,
 )
+from ray.serve._private.common import RunningReplicaInfo
 from ray.serve._private.deployment_scheduler import (
     DefaultDeploymentScheduler,
     DeploymentScheduler,
 )
+from ray.serve._private.replica_scheduler.replica_wrapper import ActorReplicaWrapper
 from ray.serve._private.utils import get_head_node_id
 
 # NOTE: Please read carefully before changing!
@@ -35,3 +37,7 @@ def create_deployment_scheduler(
         create_placement_group_fn=create_placement_group_fn_override
         or ray.util.placement_group,
     )
+
+
+def create_replica_wrapper(replica_info: RunningReplicaInfo):
+    return ActorReplicaWrapper(replica_info)
