+++ python/ray/serve/_private/replica_scheduler/__init__.py
@@ -1,8 +1,10 @@
-from ray.serve._private.replica_scheduler.common import (  # noqa: F401
-    PendingRequest,
-    ReplicaScheduler,
-    ReplicaWrapper,
-)
+from ray.serve._private.replica_scheduler.common import PendingRequest  # noqa: F401
 from ray.serve._private.replica_scheduler.pow_2_scheduler import (  # noqa: F401
     PowerOfTwoChoicesReplicaScheduler,
+)
+from ray.serve._private.replica_scheduler.replica_scheduler import (  # noqa: F401
+    ReplicaScheduler,
+)
+from ray.serve._private.replica_scheduler.replica_wrapper import (  # noqa: F401
+    ReplicaWrapper,
 )
